Move Recording class to proper package
