+++ stream-chat-android-ui-common/src/main/kotlin/io/getstream/sdk/chat/audio/recording/DefaultStreamMediaRecorder.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.getstream.sdk.chat.audio.recording
+package io.getstream.sdk.chat.audio.recording
 
 import android.content.Context
 import android.media.MediaMetadataRetriever
